chore: rename overlay component
