+++ src/components/cli-tutor-mode/CliTutorMode.js
@@ -6,7 +6,7 @@ import { connect } from 'redux-bundler-react'
 import { Modal, ModalBody, ModalActions } from '../modal/modal'
 import StrokeCode from '../../icons/StrokeCode.js'
 import Button from '../button/button.tsx'
-import Overlay from '../overlay/overlay-component'
+import Overlay from '../overlay/overlay'
 import Shell from '../shell/Shell.js'
 import StrokeDownload from '../../icons/StrokeDownload.js'
 import { cliCmdKeys, cliCommandList, cliCmdPrefixes } from '../../bundles/files/consts.js'
